pyproject.toml
[project]
name = "hotel-reservation-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.3",
    "djangorestframework>=3.17.1",
    "djangorestframework-stubs>=3.16.8",
    "drf-spectacular>=0.29.0",
    "Faker>=37.6.0",
    "gunicorn>=25.2.0",
    "mysqlclient>=2.2.8",
    "ruff>=0.15.7",
    "ty>=0.0.25",
    "uvicorn>=0.42.0",
]

